src/modules/account/useCases/showUser/ShowSignedUserUseCase.ts
import { inject, injectable } from "tsyringe";

import { IUsersRepository } from "@modules/account/repositories/IUsersRepository";
import { User } from "@modules/account/infra/typeorm/entities/User";
import { AppError } from "@shared/errors/AppError";

@injectable()
class ShowSignedUserUseCase {
  constructor(
    @inject("UsersRepository")
    private usersRepository: IUsersRepository
  ) { }

  async execute(id: string): Promise<User> {
    const user = await this.usersRepository.findById(id);

    if (!user) {
      throw new AppError("User not found", 404);
    }

    return user;
  }
}

export { ShowSignedUserUseCase };
